pyproject.toml
[tool.poetry]
name = "shablbot"
version = "1.0.3"
description = "SnablBot is an easily customizable bot capable of covering your tasks."
authors = ["Alexander Dr. <sasna142@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
vk-api = "^11.9.4"
requests = "^2.26.0"
pytz = "^2021.1"
loguru = "^0.5.3"
python-dotenv = "^0.19.0"
pydantic = "^1.8.2"
anytree = "^2.8.0"
certifi = "^2021.10.8"
charset-normalizer = "^2.0.7"
colorama = "^0.4.4"
idna = "^3.3"
six = "^1.16.0"
typing-extensions = "^3.10.0"
urllib3 = "^1.26.7"
win32-setctime = "^1.0.3"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.7"
black = "^21.9b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
